WebServer/requirements_20230306.txt
backports.functools-lru-cache==1.6.4
backports.ssl-match-hostname==3.4.0.2
certifi==2015.4.28
cycler==0.10.0
decorator==3.4.0
Django==1.7.4
django-chartit==0.1
gnureadline==6.3.3
image==1.3.9
ipython==3.1.0
Jinja2==2.7.3
jsonschema==2.4.0
kiwisolver==1.1.0
MarkupSafe==0.23
matplotlib==2.2.5
MySQL-python==1.2.5
networkx==1.9.1
nose==1.3.6
numpy==1.9.2
pandas==0.16.1
Pillow==6.2.2
pkg-resources==0.0.0
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
pyzmq==14.6.0
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.15.1
seaborn==0.5.1
six==1.10.0
subprocess32==3.5.4
sympy==0.7.6
tornado==4.1
